backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
joblib
pandas
scikit-learn
xgboost
